test fails moving template to variable
